Use itemsSummaryString64()
svn path=/trunk/kdebase/libkonq/; revision=92920

+++ lib/konq/konq_dirpart.h
@@ -239,7 +239,7 @@ protected:
 
     int m_iIconSize[4];
 
-    unsigned long m_lDirSize;
+    long long m_lDirSize;
     uint m_lFileCount;
     uint m_lDirCount;
     //bool m_bMultipleItemsSelected;
